Create README.md
Culinary arts are the cuisine arts of food preparation, cooking, and presentation of food, usually in the form of meals. People working in this field – especially in establishments such as restaurants – are commonly called "chefs" or "cooks", although, at its most general, the terms "culinary artist" and "culinarian" are also used. Table manners ("the table arts") are sometimes referred to as a culinary art.

Expert chefs are required to have knowledge of food science, nutrition and diet and are responsible for preparing meals that are as pleasing to the eye as they are to the palate. After restaurants, their primary places of work include delicatessens and relatively large institutions such as hotels and hospitals.